+++ src/main/java/com/example/labjee/services/MovieCountryService.java
@@ -1,15 +1,19 @@
 package com.example.labjee.services;
 
+import com.example.labjee.helpers.visitor.ServiceElement;
+import com.example.labjee.helpers.visitor.Visitor;
+import com.example.labjee.interfaces.MovieRelationship;
 import com.example.labjee.models.Country;
 import com.example.labjee.models.Movie;
 import com.example.labjee.models.MovieCountry;
+import com.example.labjee.models.MovieDirector;
 import com.example.labjee.repositories.MovieCountryRepository;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class MovieCountryService {
+public class MovieCountryService implements ServiceElement {
     @Autowired
     private MovieCountryRepository movieCountryRepository;
 
@@ -61,21 +65,34 @@ public class MovieCountryService {
         return 2;
     }
 
-    public int deleteLink(MovieCountry movieCountry) {
-        if (this.getLink(movieCountry.getMovie().getId(), movieCountry.getCountry().getCode()) != null) {
-            if (movieService.getById(movieCountry.getMovie().getId()) != null && countryService.getByCode(movieCountry.getCountry().getCode()) != null) {
-                Country country = countryService.getByCode(movieCountry.getCountry().getCode());
-                country.deleteMovie(movieCountry);
-                countryService.createOrUpdate(country);
+    public int deleteLink(MovieRelationship movieRelationship) {
+        if (movieRelationship instanceof MovieCountry movieCountry) {
+            if (this.getLink(movieCountry.getMovie().getId(), movieCountry.getCountry().getCode()) != null) {
+                if (movieService.getById(movieCountry.getMovie().getId()) != null && countryService.getByCode(movieCountry.getCountry().getCode()) != null) {
+                    Country country = countryService.getByCode(movieCountry.getCountry().getCode());
+                    country.deleteMovie(movieCountry);
+                    countryService.createOrUpdate(country);
 
-                movieCountryRepository.delete(movieCountry);
+                    movieCountryRepository.delete(movieCountry);
 
-                return 0;
+                    return 0;
+                }
+
+                return 1;
             }
 
-            return 1;
+            return 2;
         }
+        return 3;
+    }
 
-        return 2;
+    @Override
+    public void accept(Visitor visitor) {
+
+    }
+
+    @Override
+    public void accept(Visitor visitor, MovieRelationship relationship) {
+        visitor.visit(this, relationship);
     }
 }
