+++ src/main/java/com/example/labjee/services/MovieDirectorService.java
@@ -1,6 +1,10 @@
 package com.example.labjee.services;
 
+import com.example.labjee.helpers.visitor.ServiceElement;
+import com.example.labjee.helpers.visitor.Visitor;
+import com.example.labjee.interfaces.MovieRelationship;
 import com.example.labjee.models.Movie;
+import com.example.labjee.models.MovieActor;
 import com.example.labjee.models.MovieDirector;
 import com.example.labjee.models.Person;
 import com.example.labjee.repositories.MovieDirectorRepository;
@@ -9,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class MovieDirectorService {
+public class MovieDirectorService implements ServiceElement {
 
     @Autowired
     private MovieDirectorRepository movieDirectorRepository;
@@ -62,21 +66,34 @@ public class MovieDirectorService {
         return 2;
     }
 
-    public int deleteLink(MovieDirector movieDirector) {
-        if (this.getLink(movieDirector.getMovie().getId(), movieDirector.getDirector().getId()) != null) {
-            if (movieService.getById(movieDirector.getMovie().getId()) != null && personService.getById(movieDirector.getDirector().getId()) != null) {
-                Person person = personService.getById(movieDirector.getDirector().getId());
-                person.deleteMovieDirector(movieDirector);
-                personService.createOrUpdate(person);
+    public int deleteLink(MovieRelationship movieRelationship) {
+        if (movieRelationship instanceof MovieDirector movieDirector) {
+            if (this.getLink(movieDirector.getMovie().getId(), movieDirector.getDirector().getId()) != null) {
+                if (movieService.getById(movieDirector.getMovie().getId()) != null && personService.getById(movieDirector.getDirector().getId()) != null) {
+                    Person person = personService.getById(movieDirector.getDirector().getId());
+                    person.deleteMovieDirector(movieDirector);
+                    personService.createOrUpdate(person);
 
-                movieDirectorRepository.delete(movieDirector);
+                    movieDirectorRepository.delete(movieDirector);
 
-                return 0;
+                    return 0;
+                }
+
+                return 1;
             }
 
-            return 1;
+            return 2;
         }
+        return 3;
+    }
 
-        return 2;
+    @Override
+    public void accept(Visitor visitor) {
+
+    }
+
+    @Override
+    public void accept(Visitor visitor, MovieRelationship relationship) {
+        visitor.visit(this, relationship);
     }
 }
