+++ src/main/java/com/example/labjee/services/MovieWriterService.java
@@ -1,6 +1,10 @@
 package com.example.labjee.services;
 
+import com.example.labjee.helpers.visitor.ServiceElement;
+import com.example.labjee.helpers.visitor.Visitor;
+import com.example.labjee.interfaces.MovieRelationship;
 import com.example.labjee.models.Movie;
+import com.example.labjee.models.MovieGenre;
 import com.example.labjee.models.MovieWriter;
 import com.example.labjee.models.Person;
 import com.example.labjee.repositories.MovieWriterRepository;
@@ -9,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class MovieWriterService {
+public class MovieWriterService implements ServiceElement {
     @Autowired
     private MovieWriterRepository movieWriterRepository;
     
@@ -61,21 +65,35 @@ public class MovieWriterService {
         return 2;
     }
     
-    public int deleteLink(MovieWriter movieWriter) {
-        if (this.getLink(movieWriter.getMovie().getId(), movieWriter.getWriter().getId()) != null) {
-            if (movieService.getById(movieWriter.getMovie().getId()) != null && personService.getById(movieWriter.getWriter().getId()) != null) {
-                Person person = personService.getById(movieWriter.getWriter().getId());
-                person.deleteMovieWriter(movieWriter);
-                personService.createOrUpdate(person);
+    public int deleteLink(MovieRelationship movieRelationship) {
+        if (movieRelationship instanceof MovieWriter movieWriter) {
+            if (this.getLink(movieWriter.getMovie().getId(), movieWriter.getWriter().getId()) != null) {
+                if (movieService.getById(movieWriter.getMovie().getId()) != null && personService.getById(movieWriter.getWriter().getId()) != null) {
+                    Person person = personService.getById(movieWriter.getWriter().getId());
+                    person.deleteMovieWriter(movieWriter);
+                    personService.createOrUpdate(person);
 
-                movieWriterRepository.delete(movieWriter);
+                    movieWriterRepository.delete(movieWriter);
 
-                return 0;
+                    return 0;
+                }
+
+                return 1;
             }
 
-            return 1;
+            return 2;
         }
 
-        return 2;
+        return 3;
+    }
+
+    @Override
+    public void accept(Visitor visitor) {
+
+    }
+
+    @Override
+    public void accept(Visitor visitor, MovieRelationship relationship) {
+        visitor.visit(this, relationship);
     }
 }
